removed vite and made static and vanilla
