+++ src/imports.html
@@ -1,0 +1,2 @@
+<script src="https://cdn.jsdelivr.net/npm/apexcharts"></script>
+<script src="https://cdn.tailwindcss.com"></script>
