+++ src/index.html
@@ -1,4 +1,3 @@
-<!DOCTYPE html>
 <html lang="en">
 
 <head>
@@ -8,7 +7,7 @@
 	<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 	<script src="https://cdn.tailwindcss.com"></script>
 	<link rel="stylesheet" href="https://unpkg.com/flowbite@1.5.3/dist/flowbite.min.css" />
-
+	<script src="https://cdn.jsdelivr.net/npm/apexcharts" type="module"></script>
 	<title>FalconVis</title>
 </head>
 
@@ -26,17 +25,17 @@
 		  <div class="hidden w-full md:block md:w-auto" id="navbar-default">
 			<ul class="flex flex-col p-4 mt-2 rounded-lg border md:flex-row md:space-x-8 md:mt-0 md:text-sm md:font-medium md:border-0">
 			  <li>
-                <a href="/match.html" class="block pr-4 pl-3 text-lg" aria-current="page">Match</a>
-              </li>
-              <li>
-                <a href="/alliance.html" class="block pr-4 pl-3 text-lg" aria-current="page">Alliance</a>
-              </li>
-                <li>
-				<a href="/index.html" class="block pr-4 pl-3 text-lg" aria-current="page">Event</a>
-			  </li>
-			  <li>
-				<a href="/team.html" class="block pr-4 pl-3 text-lg" aria-current="page">Teams</a>
-			  </li>
+          <a href="/match/" class="block pr-4 pl-3 text-lg" aria-current="page">Match</a>
+        </li>
+        <li>
+          <a href="/alliance/" class="block pr-4 pl-3 text-lg" aria-current="page">Alliance</a>
+        </li>
+        <li>
+          <a href="/" class="block pr-4 pl-3 text-lg" aria-current="page">Event</a>
+        </li>
+        <li>
+          <a href="/team/" class="block pr-4 pl-3 text-lg" aria-current="page">Teams</a>
+        </li>
 			</ul>
 		  </div>
 		</div>
