+++ src/lib/components/BarGraph.js
@@ -1,4 +1,4 @@
-import { Graph } from "./Graph"
+import { Graph } from "./Graph.js"
 
 class BarGraph {
     constructor(parent_id, title, plotOptions, dataOptions, modal, editable = true) {
