+++ src/lib/components/StatNumber.js
