+++ src/match/index.html
@@ -1,4 +1,3 @@
-<!DOCTYPE html>
 <html lang="en">
 
 <head>
@@ -8,7 +7,7 @@
 	<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 	<script src="https://cdn.tailwindcss.com"></script>
 	<link rel="stylesheet" href="https://unpkg.com/flowbite@1.5.3/dist/flowbite.min.css" />
-
+	<script src="https://cdn.jsdelivr.net/npm/apexcharts" type="module"></script>
 	<title>FalconVis</title>
 </head>
 
@@ -16,7 +15,7 @@
 	<nav class="bg-white border-gray-300 px-2 sm:px-4 py-2.5 border-b-2 mb-4">
 		<div class="container flex flex-wrap justify-between items-center mx-auto">
 		  <a href="/" class="flex items-center">
-			  <img src="public/logo.svg" class="mr-3 h-6 sm:h-9" alt="Flowbite Logo" />
+			  <img src="../public/logo.svg" class="mr-3 h-6 sm:h-9" alt="Flowbite Logo" />
 			  <span class="self-center text-2xl font-semibold whitespace-nowrap">FalconVis</span>
 		  </a>
 		  <button data-collapse-toggle="navbar-default" type="button" class="inline-flex items-center p-2 ml-3 text-sm text-gray-500 rounded-lg md:hidden hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-gray-200 dark:text-gray-400 dark:hover:bg-gray-700 dark:focus:ring-gray-600" aria-controls="navbar-default" aria-expanded="false">
@@ -26,17 +25,17 @@
 		  <div class="hidden w-full md:block md:w-auto" id="navbar-default">
 			<ul class="flex flex-col p-4 mt-2 rounded-lg border md:flex-row md:space-x-8 md:mt-0 md:text-sm md:font-medium md:border-0">
 			  <li>
-                <a href="/match.html" class="block pr-4 pl-3 text-lg" aria-current="page">Match</a>
-              </li>
-              <li>
-                <a href="/alliance.html" class="block pr-4 pl-3 text-lg" aria-current="page">Alliance</a>
-              </li>
-                <li>
-				<a href="/index.html" class="block pr-4 pl-3 text-lg" aria-current="page">Event</a>
-			  </li>
-			  <li>
-				<a href="/team.html" class="block pr-4 pl-3 text-lg" aria-current="page">Teams</a>
-			  </li>
+          <a href="/match/" class="block pr-4 pl-3 text-lg" aria-current="page">Match</a>
+        </li>
+        <li>
+          <a href="/alliance/" class="block pr-4 pl-3 text-lg" aria-current="page">Alliance</a>
+        </li>
+        <li>
+          <a href="/" class="block pr-4 pl-3 text-lg" aria-current="page">Event</a>
+        </li>
+        <li>
+          <a href="/team/" class="block pr-4 pl-3 text-lg" aria-current="page">Teams</a>
+        </li>
 			</ul>
 		  </div>
 		</div>
@@ -51,29 +50,29 @@
 			<div class="flex flex-wrap gap-4 w-[400px]">
 				<div class="">
 					<h1 class="text-lg font-medium text-blue-700">Blue 1</h1>
-					<input id="blue1" class="h-8 border w-28 border-gray-400 rounded-md pl-1" placeholder="1577" type="number"></input>
+					<input id="blue1" class="h-8 border w-28 border-gray-400 rounded-md pl-1" placeholder="1577" type="number"/>
 				</div>
 				<div class="">
 					<h1 class="text-lg font-medium text-blue-700">Blue 2</h1>
-					<input id="blue2" class="h-8 border w-28 border-gray-400 rounded-md pl-1" placeholder="4414" type="number"></input>
+					<input id="blue2" class="h-8 border w-28 border-gray-400 rounded-md pl-1" placeholder="4414" type="number"/>
 				</div>
 				<div class="">
 					<h1 class="text-lg font-medium text-blue-700">Blue 3</h1>
-					<input id="blue3" class="h-8 border w-28 border-gray-400 rounded-md pl-1" placeholder="4099" type="number"></input>
+					<input id="blue3" class="h-8 border w-28 border-gray-400 rounded-md pl-1" placeholder="4099" type="number"/>
 				</div>
 			</div>
 			<div class="flex flex-wrap gap-4 w-[400px]">
 				<div class="">
 					<h1 class="text-lg font-medium text-red-500">Red 1</h1>
-					<input id="red1" class="h-8 border w-28 border-gray-400 rounded-md pl-1" placeholder="1577" type="number"></input>
+					<input id="red1" class="h-8 border w-28 border-gray-400 rounded-md pl-1" placeholder="1577" type="number"/>
 				</div>
 				<div class="">
 					<h1 class="text-lg font-medium text-red-500">Red 2</h1>
-					<input id="red2" class="h-8 border w-28 border-gray-400 rounded-md pl-1" placeholder="4414" type="number"></input>
+					<input id="red2" class="h-8 border w-28 border-gray-400 rounded-md pl-1" placeholder="4414" type="number"/>
 				</div>
 				<div class="">
 					<h1 class="text-lg font-medium text-red-500">Red 3</h1>
-					<input id="red3" class="block h-8 border w-28 border-gray-400 rounded-md pl-1" placeholder="4099" type="number"></input>
+					<input id="red3" class="block h-8 border w-28 border-gray-400 rounded-md pl-1" placeholder="4099" type="number"/>
 					<button id="setTeams" class="h-10 w-28 rounded-md bg-[#efae04] text-white font-semibold mt-2">Submit</button>
 				</div>
 			</div>
