+++ src/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "http-server": "^14.1.1"
+  }
+}
