+++ src/team/index.html
@@ -8,7 +8,7 @@
 	<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 	<script src="https://cdn.tailwindcss.com"></script>
 	<link rel="stylesheet" href="https://unpkg.com/flowbite@1.5.3/dist/flowbite.min.css" />
-
+	<script src="https://cdn.jsdelivr.net/npm/apexcharts" type="module"></script>
 	<title>FalconVis</title>
 </head>
 
@@ -16,7 +16,7 @@
 	<nav class="bg-white border-gray-300 px-2 sm:px-4 py-2.5 border-b-2 mb-4">
 		<div class="container flex flex-wrap justify-between items-center mx-auto">
 		  <a href="/" class="flex items-center">
-			  <img src="public/logo.svg" class="mr-3 h-6 sm:h-9" alt="Flowbite Logo" />
+			  <img src="../public/logo.svg" class="mr-3 h-6 sm:h-9" alt="Flowbite Logo" />
 			  <span class="self-center text-2xl font-semibold whitespace-nowrap">FalconVis</span>
 		  </a>
 		  <button data-collapse-toggle="navbar-default" type="button" class="inline-flex items-center p-2 ml-3 text-sm text-gray-500 rounded-lg md:hidden hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-gray-200 dark:text-gray-400 dark:hover:bg-gray-700 dark:focus:ring-gray-600" aria-controls="navbar-default" aria-expanded="false">
@@ -26,17 +26,17 @@
 		  <div class="hidden w-full md:block md:w-auto" id="navbar-default">
 			<ul class="flex flex-col p-4 mt-2 rounded-lg border md:flex-row md:space-x-8 md:mt-0 md:text-sm md:font-medium md:border-0">
 			  <li>
-                <a href="/match.html" class="block pr-4 pl-3 text-lg" aria-current="page">Match</a>
-              </li>
-              <li>
-                <a href="/alliance.html" class="block pr-4 pl-3 text-lg" aria-current="page">Alliance</a>
-              </li>
-                <li>
-				<a href="/index.html" class="block pr-4 pl-3 text-lg" aria-current="page">Event</a>
-			  </li>
-			  <li>
-				<a href="/team.html" class="block pr-4 pl-3 text-lg" aria-current="page">Teams</a>
-			  </li>
+          <a href="/match/" class="block pr-4 pl-3 text-lg" aria-current="page">Match</a>
+        </li>
+        <li>
+          <a href="/alliance/" class="block pr-4 pl-3 text-lg" aria-current="page">Alliance</a>
+        </li>
+        <li>
+          <a href="/" class="block pr-4 pl-3 text-lg" aria-current="page">Event</a>
+        </li>
+        <li>
+          <a href="/team/" class="block pr-4 pl-3 text-lg" aria-current="page">Teams</a>
+        </li>
 			</ul>
 		  </div>
 		</div>
