--- vite.config.js
@@ -1,6 +1,0 @@
-import { defineConfig } from 'vite'
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  publicDir: 'src'
-})
